Initial Refactor refactor
